refactor (Racing.java): service 패키지로 이동
