+++ src/test/java/racinggame/service/RacingTest.java
@@ -1,4 +1,4 @@
-package racinggame.domain.racing;
+package racinggame.service;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -6,6 +6,7 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import racinggame.servcie.Racing;
 
 import java.util.ArrayList;
 import java.util.Arrays;
